feat: added to root

+++ normalize.go
@@ -3,14 +3,14 @@ package hnsw
 import "math"
 
 // normalizing for faster cosineDistance
-func Normalize(vec []float64) {
-	var norm float64
+func Normalize(vec []float32) {
+	var norm float32
 
 	for _, x := range vec {
 		norm += x * x
 	}
 
-	norm = math.Sqrt(norm)
+	norm = float32(math.Sqrt(float64(norm)))
 
 	if norm == 0 {
 		return
